moved install logic to install submodule
