+++ tests/unit_tests/cli/hub/test_install.py
@@ -9,9 +9,7 @@ from guardrails.cli.server.module_manifest import ModuleManifest
 
 class TestInstall:
     def test_exits_early_if_uri_is_not_valid(self, mocker):
-        mock_logger_error = mocker.patch(
-            "guardrails.hub.validator_package_service.cli_logger.error"
-        )
+        mock_logger_error = mocker.patch("guardrails.hub.install.cli_logger.error")
 
         runner = CliRunner()
         result = runner.invoke(hub_command, ["install", "some-invalid-uri"])
@@ -22,9 +20,7 @@ class TestInstall:
         )
 
     def test_install_local_models__false(self, mocker):
-        mock_install = mocker.patch(
-            "guardrails.hub.validator_package_service.ValidatorPackageService.install"
-        )
+        mock_install = mocker.patch("guardrails.hub.install.install")
         runner = CliRunner()
         result = runner.invoke(
             hub_command,
@@ -41,9 +37,7 @@ class TestInstall:
         assert result.exit_code == 0
 
     def test_install_local_models__true(self, mocker):
-        mock_install = mocker.patch(
-            "guardrails.hub.validator_package_service.ValidatorPackageService.install"
-        )
+        mock_install = mocker.patch("guardrails.hub.install.install")
         runner = CliRunner()
         result = runner.invoke(
             hub_command,
@@ -59,9 +53,7 @@ class TestInstall:
         assert result.exit_code == 0
 
     def test_install_local_models__none(self, mocker):
-        mock_install = mocker.patch(
-            "guardrails.hub.validator_package_service.ValidatorPackageService.install"
-        )
+        mock_install = mocker.patch("guardrails.hub.install.install")
         runner = CliRunner()
         result = runner.invoke(
             hub_command,
@@ -77,9 +69,7 @@ class TestInstall:
         assert result.exit_code == 0
 
     def test_install_quiet(self, mocker):
-        mock_install = mocker.patch(
-            "guardrails.hub.validator_package_service.ValidatorPackageService.install"
-        )
+        mock_install = mocker.patch("guardrails.hub.install.install")
         runner = CliRunner()
         result = runner.invoke(
             hub_command, ["install", "hub://guardrails/test-validator", "--quiet"]
